+++ src/components/InfoSection.js
@@ -1,17 +1,19 @@
-import React, {PureComponent} from 'react';
-import ReactMarkdown from 'react-markdown';
+import React, { PureComponent } from "react";
+import ReactMarkdown from "react-markdown";
 
-import md from '../markdown/InfoSection.md';
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faInfoCircle } from "@fortawesome/free-solid-svg-icons";
+
+import md from "../markdown/InfoSection.md";
 
 /**
  * The InfoSection which resides in the InfoModal.
  * Renders markdown from markdown/InfoSection.md.
  */
 class InfoSection extends PureComponent {
-
   constructor() {
     super();
-    this.state = { markdown: '' };
+    this.state = { markdown: "" };
   }
 
   /**
@@ -19,7 +21,9 @@ class InfoSection extends PureComponent {
    * @return {void}
    */
   componentWillMount() {
-    fetch(md).then(res => res.text()).then(text => this.setState({ markdown: text }));
+    fetch(md)
+      .then((res) => res.text())
+      .then((text) => this.setState({ markdown: text }));
   }
 
   /**
@@ -29,10 +33,17 @@ class InfoSection extends PureComponent {
   render() {
     return (
       <section>
-        <ReactMarkdown>{this.state.markdown}</ReactMarkdown>
+        <div className="note">
+          <FontAwesomeIcon icon={faInfoCircle} />
+          Please note that the Planner is not fully updated to V2.0 yet. New
+          creatures/traits have been added, but they are missing sprites.
+          Specializations and annointments have not been added yet - they will
+          be added once the data becomes available.
+        </div>
+        <ReactMarkdown safe={true}>{this.state.markdown}</ReactMarkdown>
       </section>
-      )
-    }
+    );
+  }
 }
 
 export default InfoSection;